Update to version 3.3f:
Fixed problem with \longmapsto that caused gaps between the characters
that form the arrow.  Also worked around a problem with Firefox that
caused \mapsto to lose its vertical tick at the left-hand end of the
arrow.

Fixed a problem with autoload where multiple calls to tex2math could
cause tex2math to convert \$ to $ and then later try to interpret the
$ as a start-of-math delimiter, even though it was previously escaped.

Reorganized the autoload plugin so that jsMath.Autoload.Run() no
longer needs to be called.  In the past, autoload would use tex2math
to look for text that needs to be converted to SPAN and DIV tags of
class "math", but would not convert them itself.  Now that autoload
has settings to control what tex2math functions will be performed,
there is no need to put off the conversion, so autoload now does the
tex2math conversion immediately.  The jsMath.Autoload.findMathElements
variable is no longer needed, and so has been depricated.

Fixed a problem with tex2math where two escaped dollars in a row would
not be processed properly (the second would not be treated as escaped,
and so could be treated as a start-of-math delimiter).