Changes for 3.4a:
Fixed a problem with the noImageFonts plugin where extra
fonts would still try to use image fonts even when they were supposed
to be disabled.  Also, noImageFonts now adjusts the error message to
indicate that unicode fonts will be used for fallback rather than
image fonts.

Fixed a problem where changes in the CSS line-height attribute could
cause alignment problems with arrays in MSIE.

Fixed a number of alignment problems with MSIE when used with a
DOCTYPE that triggers "standards" mode rather than "quirks" mode.

A bug fix needed for MSIE6 that is one of the sources of jsMath's
slowness with MSIE is now disabled when MSIE7 is used, since it is no
longer needed in that version.  This should give a noticable speed
improvement for jsMath in MSIE7.