Bring jsMath to version 3.4e
* Added ability for the tex2math to find LaTeX environments of the
form \begin{xxx}...\end{xxx} outside of math mode.  This is enabled by
default for jsMath.ConvertLaTeX(), and can be enabled in the
easy/load.js by setting the processLaTeXenvironments value to 1 (now
the default).

* Fixed issues with loading a local copy of jsMath in
MSIE. Apparently, MSIE considers xmlHttpRequest calls for local files
to be access violations, even when the calling file is also local, and
so MSIE needs to use the non-xmlHttpRequest fallback method when
loading jsMath locally.

* Worked around a long-standing issue with MSIE where elements that
are positioned using backwards skips could be clipped in some
circumstances.  In particular, the Chapter 18 commutative diagram now
displays correctly, as do exponents that have been adjusted to the
left.  This was a vexing MSIE bug that is good to finally have worked
around.

* Fixed a problem with placement of accents in some situations (the
italic correction was not being applied properly).

* Worked around bug in MSIE where the \vec vector arrow was being
clipped at the right.

* Worked around bug in MSIE where \mapsto and \longmapsto were
displayed incorrectly (space appeared between the arrow and the tail).