Bring to version 3.5.  Changes include:
* Fixed a problem with the boldsymbol extension where it could leave
the fonts set to the bold fonts, so that all the rest of the
mathematics on the page would be in bold.  (This would occur if a
jsMath auxiliary file had to be autoloaded by the \boldsymbol
command.)

* Fixed a timing problem with easy/load.js when autoloading and
tex2math are both used (the default) that could cause some internal
routines to run in the wrong order.

* Fixed a problem in detecting a font size change in MSIE when the
font used at one location on the page is only slightly smaller or
larger than that used at the previous location.

* Added ability for extra fonts to specify how they are handled in
symbols-only image mode.  This is used by the cmmib10 font to specify
that the italic letters should be handled by bold italic letters in
the serif font rather than by images, which was the old behaviour.
You will need to replace the jsMath/fonts/cmmib10/def.js file with the
new one to take advantage of this.

* Added a new autobold extension that allows jsMath to detect the
presence of a <B>...</B> tags (or CSS that sets font-weight:bold)
surrounding the mathematics and automatically apply \boldsymbol to the
mathematics.  Since this requires the extra cmmib10 and cmmsy10 fonts,
this is an extension that you must activate yourself.  See the
autobold documentation for details

* Implemented the \verb command for entering verbatim material.  See
the verb extension documentation for details.

* Added aligned and gathered environments to the AMSmath plugin.

* Added equation and equation* environments.  (Since tex2math now
processes LaTeX environments within the text of the page, it makes
more sense to have these defined.)

* Fixed a problem with \newcommand where it would reject
control-sequence names that included upper-case letters.

* Fixed a problem with the \font command where the control sequence
would be lost when the font was loaded (if it wasn't already loaded
when the \font command was issued).

* Fixed a timing bug with easy/load.js when there are files specified
in the loadFiles array.  The bug only occurred with the Opera browser.

* Worked around a problem with Opera where the autoload plugin could
hang when trying to load the jsMath.js file.

* Worked around a spacing problem in IE7 where some elements in an
array or fraction could be misaligned.

* Worked around a problem in IE7 where accented characters could
disappear under some circumstances (particularly when they appear by
themselves).

* Fixed a problem with Firefox3 where the control panel would not
open.

* Firefox3's new single-origin policy restricts access to file outside
the currect directory when local files (i.e., file: URL's) are used.
JsMath now uses the new window.postMessage() facilty to work around
this.  If you use window.postMessage() yourself you should be aware
that jsMath is also using window messages, and so you should be
careful to properly identify your own messages so as not to confuse
them with the ones used by jsMath.  All messages used by jsMath will
begin with jsMLD:, jsMCP:, or jsMAL: so you can filter these out of
our own message-processing code.

* Disabled the BaKoMa font processing in Firefox3 since jsMath no
longer works properly with those fonts in Firefox3.