Updated to version 3.6a.  Changes include:
Worked around a problem with image fonts in Firefox3. The width of the
images were not being properly reported (even when specified
Killed by signal 2.MG tag's STYLE attribute) if the image wasn't
loaded from the server yet.

Worked around a problem in IE8 when jsMath is loaded from the local
disk rather than from a server.

Added more checks when jsMath.safeHBoxes is set to prevent additional
ways in which the user could have inserted HTML tags into \hbox
commands.

Fixed a bug in the spriteImageFonts plugin that had been causing it to
fail in Firefox2 and earlier, Opera 9.2x and earlier, and all versions
of MSIE. It seems to have stopped working somewhere after version 3.4
of jsMath.

Fixed a bug where newlines were not treated as spaces when scanning
for arguments of macros that require arguments.

Fixed the eqn-number extension to handle \nonumber as a synonym for
\nolabel, and made the equation*, align*, and multline* environments
not produce equation numbers.

Removed a debugging alert from the Firefox3 local-file code that was
accidentally left active in version 3.6.